Moved FW-KENSC files into their own folder
Also made them ignore the precompiled header nonsense. It's much cleaner
now.
