+++ FW-KENSC-ShellExt/FW-KENSC/basic_decoder.h
