+++ FW-KENSC-ShellExt/FW-KENSC/bigendian_io.h
