+++ FW-KENSC-ShellExt/FW-KENSC/bitstream.h
