+++ FW-KENSC-ShellExt/FW-KENSC/comper.cc
@@ -1,6 +1,3 @@
-#include "StdAfx.h"
-#undef min
-#undef max
 /* -*- Mode: C++; indent-tabs-mode: t; c-basic-offset: 4; tab-width: 4 -*- */
 /*
  * Copyright (C) Flamewing 2013-2016 <flamewing.sonic@gmail.com>
