+++ FW-KENSC-ShellExt/FW-KENSC/comper.h
