+++ FW-KENSC-ShellExt/FW-KENSC/enigma.cc
@@ -1,4 +1,3 @@
-#include "StdAfx.h"
 /* -*- Mode: C++; indent-tabs-mode: t; c-basic-offset: 4; tab-width: 4 -*- */
 /*
  * Copyright (C) Flamewing 2011-2016 <flamewing.sonic@gmail.com>
