+++ FW-KENSC-ShellExt/FW-KENSC/enigma.h
