+++ FW-KENSC-ShellExt/FW-KENSC/kosinski.h
