+++ FW-KENSC-ShellExt/FW-KENSC/kosplus.h
