+++ FW-KENSC-ShellExt/FW-KENSC/moduled_adaptor.h
