+++ FW-KENSC-ShellExt/FW-KENSC/nemesis.cc
@@ -1,6 +1,3 @@
-#include "StdAfx.h"
-#undef min
-#undef max
 /* -*- Mode: C++; indent-tabs-mode: t; c-basic-offset: 4; tab-width: 4 -*- */
 /*
  * Copyright (C) Flamewing 2011-2016 <flamewing.sonic@gmail.com>
