+++ FW-KENSC-ShellExt/FW-KENSC/nemesis.h
